clean (FirmwareInstaller.cs): split up the file into partials for the sake of readability
