+++ Laerdal.McuMgr/Laerdal.McuMgr.csproj
@@ -178,6 +178,14 @@
         <Compile Update="**\FileDownloader.Commandable.*.cs">
             <DependentUpon>FileDownloader.Commandable.cs</DependentUpon>
         </Compile>
+
+        <!-- FIRMWARE INSTALLER -->
+        <Compile Update="**\FirmwareInstaller.*.cs">
+            <DependentUpon>FirmwareInstaller.cs</DependentUpon>
+        </Compile>
+        <Compile Update="**\FirmwareInstaller.Commandable.*.cs">
+            <DependentUpon>FirmwareInstaller.Commandable.cs</DependentUpon>
+        </Compile>
     </ItemGroup>
 
     <!-- ANDROID -->
