+++ Laerdal.McuMgr/Shared/FirmwareInstallation/FirmwareInstaller.Cleanupable.TryCleanupResourcesOfLastInstallation.cs
@@ -1,0 +1,7 @@
+namespace Laerdal.McuMgr.FirmwareInstallation
+{
+    public partial class FirmwareInstaller
+    {
+        public void TryCleanupResourcesOfLastInstallation() => _nativeFirmwareInstallerProxy?.TryCleanupResourcesOfLastInstallation();
+    }
+}
