+++ Laerdal.McuMgr/Shared/FirmwareInstallation/FirmwareInstaller.Commandable.Cancel.cs
@@ -1,0 +1,7 @@
+namespace Laerdal.McuMgr.FirmwareInstallation
+{
+    public partial class FirmwareInstaller
+    {
+        public void Cancel() => _nativeFirmwareInstallerProxy?.Cancel();
+    }
+}
