+++ Laerdal.McuMgr/Shared/FirmwareInstallation/FirmwareInstaller.Commandable.Disconnect.cs
@@ -1,0 +1,7 @@
+namespace Laerdal.McuMgr.FirmwareInstallation
+{
+    public partial class FirmwareInstaller
+    {
+        public void Disconnect() => _nativeFirmwareInstallerProxy?.Disconnect();
+    }
+}
