+++ Laerdal.McuMgr/Shared/FirmwareInstallation/FirmwareInstaller.Commandable.cs
@@ -1,0 +1,6 @@
+namespace Laerdal.McuMgr.FirmwareInstallation
+{
+    public partial class FirmwareInstaller
+    {
+    }
+}
